refactor - separate genre model
