+++ lib/models/genre.dart
@@ -1,0 +1,7 @@
+class Genre {
+  final String name;
+  final String imgUrl;
+  final String topicCode;
+
+  Genre(this.name, this.imgUrl, this.topicCode);
+}
